+++ bin/c-kc.js
@@ -11,6 +11,7 @@ program
     .command('clean <location>', 'Clean images specific to a Docker location.')
     .command('context', 'Get and set the kubectrl context.')
     .command('deploy <location>', 'Update all instances of a Docker location to the latest tag.')
+    .command('logs', 'View all logs from Kubernetes pods.')
     .command('start', 'Deploy all Kubernetes locations.')
     .command('stop', 'Stop and remove certain Kubernetes objects described in Kubernetes locations.')
     .parse(process.argv);