+++ bin/c.js
@@ -17,6 +17,7 @@ program
     .command('mongo <command>', 'Shortcuts to help with the database.')
     .command('npm <command>', 'Shortcuts to help with NPM.')
     .command('project <command>', 'Shortcuts to help with project management.')
+    .command('workflow <command>', 'Common workflows that can be executed within the context of a project.')
     .command('yarn <command>', 'Shortcuts to help with Yarn.')
     .parse(process.argv);