+++ CHANGELOG.md
@@ -13,7 +13,7 @@ This file is a history of the changes made to @idearium/cli.
 - Added `c hosts add`, `c hosts get` and `c hosts remove` to help with easy hosts management.
 - Added `c project env ls`, `c project env get` and `c project env set`, `c project init`, `c project name`, `c project organisation`, `c project prefix` to help with project environment management.
 - Added a `c d build` command.
-- Added `c kc apply`, `c kc build`, `c kc clean`, `c kc cmd`, `c kc context`, `c kc context get`, `c kc context set`, `c kc deploy`, `c kc logs`, `c kc pod`, `c kc start`, `c kc stop` and `c kc test` commands to help with kubectrl.
+- Added `c kc apply`, `c kc build`, `c kc clean`, `c kc cmd`, `c kc context`, `c kc context get`, `c kc context set`, `c kc deploy`, `c kc logs`, `c kc pod`, `c kc secret`, `c kc start`, `c kc stop` and `c kc test` commands to help with kubectrl.
 - Added `c gc cmd` and `c gc create` commands to help with gcloud.
 - Added `c fs symlink` commands to help with the file system.
 - Added the concept of a workflow, allowing project specific needs for predefined workflows. Current workflows are `project-init` and `cli-init`.